backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-dotenv
google-genai
python-multipart
openai-whisper
edge-tts
sqlalchemy
aiofiles